refactor(channel): collect existing implementations
