+++ asyncband/src/channel/mpsc/mod.rs
@@ -18,16 +18,16 @@
 //! A multi-producer, single-consumer queue for sending values between asynchronous tasks.
 
 mod bounded;
-mod error;
 mod unbounded;
 
 pub use bounded::BoundedReceiver;
 pub use bounded::BoundedSender;
 pub use bounded::bounded;
-pub use error::RecvError;
-pub use error::SendError;
-pub use error::TryRecvError;
-pub use error::TrySendError;
 pub use unbounded::UnboundedReceiver;
 pub use unbounded::UnboundedSender;
 pub use unbounded::unbounded;
+
+pub use super::error::RecvError;
+pub use super::error::SendError;
+pub use super::error::TryRecvError;
+pub use super::error::TrySendError;
