+++ asyncband/src/channel/oneshot/receiver.rs
@@ -24,16 +24,16 @@ use std::sync::atomic::fence;
 use std::task::Context;
 use std::task::Poll;
 
-use crate::oneshot::AWAKING;
-use crate::oneshot::Channel;
-use crate::oneshot::DISCONNECTED;
-use crate::oneshot::EMPTY;
-use crate::oneshot::MESSAGE;
-use crate::oneshot::RECEIVING;
+use super::AWAKING;
+use super::Channel;
+use super::DISCONNECTED;
+use super::EMPTY;
+use super::MESSAGE;
+use super::RECEIVING;
 #[cfg(doc)]
-use crate::oneshot::Sender;
-use crate::oneshot::deallocate_empty_channel;
-use crate::oneshot::drop_message_and_deallocate_channel;
+use super::Sender;
+use super::deallocate_empty_channel;
+use super::drop_message_and_deallocate_channel;
 
 /// Receives a value from the associated [`Sender`].
 pub struct Receiver<T> {
