+++ asyncband/src/channel/oneshot/sender.rs
@@ -22,15 +22,15 @@ use std::ptr::NonNull;
 use std::sync::atomic::Ordering;
 use std::sync::atomic::fence;
 
-use crate::oneshot::Channel;
-use crate::oneshot::DISCONNECTED;
-use crate::oneshot::EMPTY;
-use crate::oneshot::MESSAGE;
-use crate::oneshot::RECEIVING;
+use super::Channel;
+use super::DISCONNECTED;
+use super::EMPTY;
+use super::MESSAGE;
+use super::RECEIVING;
 #[cfg(doc)]
-use crate::oneshot::Receiver;
-use crate::oneshot::deallocate_empty_channel;
-use crate::oneshot::drop_message_and_deallocate_channel;
+use super::Receiver;
+use super::deallocate_empty_channel;
+use super::drop_message_and_deallocate_channel;
 
 /// Sends a value to the associated [`Receiver`].
 pub struct Sender<T> {
